Improve functional test framework

+++ src/it/scala/io/github/spark_redshift_community/spark/redshift/pushdown/IntegrationPushdownSuiteBase.scala
@@ -17,23 +17,36 @@ package io.github.spark_redshift_community.spark.redshift.pushdown
 
 import io.github.spark_redshift_community.spark.redshift.{IntegrationSuiteBase, Utils}
 import io.github.spark_redshift_community.spark.redshift.Parameters.{PARAM_AUTO_PUSHDOWN, PARAM_PUSHDOWN_UNLOAD_S3_FORMAT}
-import org.apache.spark.sql.DataFrameReader
+import org.apache.spark.sql.{DataFrameReader, SQLContext}
 
 class IntegrationPushdownSuiteBase extends IntegrationSuiteBase {
-  protected val test_table: String = s""""PUBLIC"."pushdown_suite_test_table_$randomSuffix""""
+  protected var test_table: String = setTestTableName()
   protected val test_table_safe_null = s""""PUBLIC"."pushdown_suite_test_safe_null_$randomSuffix""""
+  // This flag controls whether auto pushdown operations into Redshift is enabled.
+  protected val auto_pushdown: String = "true"
+  // This flag is useful only if auto pushdown is enabled.
+  // Otherwise, the s3 format will always be default.
   protected val s3format: String = "DEFAULT"
+  // This flag controls whether to use preloaded Redshift correctness test dataset on cluster.
+  protected val preloaded_data: String = "false"
+
+  def setTestTableName(): String = s""""PUBLIC"."pushdown_suite_test_table_$randomSuffix""""
+
   override def beforeAll(): Unit = {
     super.beforeAll()
-    conn.prepareStatement(s"drop table if exists $test_table").executeUpdate()
-    conn.prepareStatement(s"drop table if exists $test_table_safe_null").executeUpdate()
-    createTestDataInRedshift(test_table)
+    if (!preloaded_data.toBoolean) {
+      conn.prepareStatement(s"drop table if exists $test_table").executeUpdate()
+      conn.prepareStatement(s"drop table if exists $test_table_safe_null").executeUpdate()
+      createTestDataInRedshift(test_table)
+    }
   }
 
   override def afterAll(): Unit = {
     try {
-      conn.prepareStatement(s"drop table if exists $test_table").executeUpdate()
-      conn.prepareStatement(s"drop table if exists $test_table_safe_null").executeUpdate()
+      if (!preloaded_data.toBoolean) {
+        conn.prepareStatement(s"drop table if exists $test_table").executeUpdate()
+        conn.prepareStatement(s"drop table if exists $test_table_safe_null").executeUpdate()
+      }
     } finally {
       super.afterAll()
     }
@@ -55,15 +68,29 @@ class IntegrationPushdownSuiteBase extends IntegrationSuiteBase {
       .option("url", jdbcUrl)
       .option("tempdir", tempDir)
       .option("forward_spark_s3_credentials", "true")
-      .option(PARAM_AUTO_PUSHDOWN, "true")
+      .option(PARAM_AUTO_PUSHDOWN, auto_pushdown)
       .option(PARAM_PUSHDOWN_UNLOAD_S3_FORMAT, s3format)
   }
 
   def checkSqlStatement(expectedAnswer: String): Unit = {
-    assert(
-      Utils.lastBuildStmt.replaceAll("\\s", "")
-      ==
-      expectedAnswer.replaceAll("\\s", "")
+    // If there is no operation pushed down into Redshift, there is no need to
+    // validate executed statement in Redshift as it will be a simple select * statement.
+    if (auto_pushdown.toBoolean) {
+      assert(
+        Utils.lastBuildStmt.replaceAll("\\s", "")
+          ==
+          expectedAnswer.replaceAll("\\s", "")
+      )
+    }
+  }
+
+  // Method that performs validation of statement result and statement executed
+  // on Redshift.
+  def doTest(sqlContext: SQLContext, tc: TestCase): Unit = {
+    checkAnswer( // ensure statement result is as expected
+      sqlContext.sql(tc.sparkStatement),
+      tc.expectedResult
     )
+    checkSqlStatement(tc.expectedPushdownStatement)
   }
 }
